requirements.txt
certifi==2019.6.16
Django==2.2.3
djangorestframework==3.10.1
pipenv==2018.11.26
pkg-resources==0.0.0
pytz==2019.1
sqlparse==0.3.0
virtualenv==16.7.2
virtualenv-clone==0.5.3
